refactor: typescript contracts
