+++ contracts/toCKB-typescript/src/switch/deposit_request/mod.rs
@@ -1,7 +1,7 @@
 use crate::switch::ToCKBCellDataTuple;
 use crate::utils::{
     config::PLEDGE,
-    tools::*,
+    transaction::*,
     types::{Error, ToCKBCellDataView},
 };
 use ckb_std::{
