+++ contracts/toCKB-typescript/src/switch/pre_undercollateral_redeem/mod.rs
@@ -1,6 +1,6 @@
 use crate::switch::ToCKBCellDataTuple;
 use crate::utils::config::{CKB_UNITS, PRE_UNDERCOLLATERAL_RATE, XT_CELL_CAPACITY};
-use crate::utils::tools::{get_price, is_XT_typescript, XChainKind};
+use crate::utils::transaction::{get_price, get_sum_sudt_amount, is_XT_typescript, XChainKind};
 use crate::utils::types::{Error, ToCKBCellDataView};
 use ckb_std::ckb_constants::Source;
 use ckb_std::debug;
@@ -10,6 +10,19 @@ use ckb_std::high_level::{
 };
 use core::result::Result;
 use molecule::prelude::*;
+
+pub fn verify(toCKB_data_tuple: &ToCKBCellDataTuple) -> Result<(), Error> {
+    let input_toCKB_data = toCKB_data_tuple
+        .0
+        .as_ref()
+        .expect("inputs contain toCKB cell");
+    let lot_size = match input_toCKB_data.get_xchain_kind() {
+        XChainKind::Btc => input_toCKB_data.get_btc_lot_size()?.get_sudt_amount(),
+        XChainKind::Eth => input_toCKB_data.get_eth_lot_size()?.get_sudt_amount(),
+    };
+    verify_collateral_rate(lot_size)?;
+    verify_burn(lot_size, input_toCKB_data)
+}
 
 fn verify_burn(lot_size: u128, data: &ToCKBCellDataView) -> Result<(), Error> {
     let lock_hash = load_cell_lock_hash(0, Source::GroupInput)?;
@@ -27,9 +40,7 @@ fn verify_burn(lot_size: u128, data: &ToCKBCellDataView) -> Result<(), Error> {
                 if lock.as_bytes() == data.signer_lockscript {
                     is_signer = true;
                 }
-                if !(cell_type.is_some()
-                    && is_XT_typescript(&cell_type.unwrap(), lock_hash.as_ref()))
-                {
+                if !is_XT_typescript(&cell_type, lock_hash.as_ref()) {
                     input_index += 1;
                     continue;
                 }
@@ -49,33 +60,9 @@ fn verify_burn(lot_size: u128, data: &ToCKBCellDataView) -> Result<(), Error> {
         return Err(Error::InputSignerInvalid);
     }
 
-    let mut output_sudt_num = 0;
-    let mut output_index = 0;
-    loop {
-        let cell_type = load_cell_type(output_index, Source::Output);
-        match cell_type {
-            Err(SysError::IndexOutOfBound) => break,
-            Err(_err) => panic!("iter output return an error"),
-            Ok(cell_type) => {
-                if !(cell_type.is_some()
-                    && is_XT_typescript(&cell_type.unwrap(), lock_hash.as_ref()))
-                {
-                    output_index += 1;
-                    continue;
-                }
+    let output_sudt_sum = get_sum_sudt_amount(0, Source::Output, lock_hash.as_ref())?;
 
-                let data = load_cell_data(output_index, Source::Output)?;
-                let mut buf = [0u8; 16];
-                if data.len() == 16 {
-                    buf.copy_from_slice(&data);
-                    output_sudt_num += u128::from_le_bytes(buf)
-                }
-                output_index += 1;
-            }
-        }
-    }
-
-    if input_sudt_sum - output_sudt_num != lot_size {
+    if input_sudt_sum - output_sudt_sum != lot_size {
         return Err(Error::XTBurnInvalid);
     }
 
@@ -97,17 +84,4 @@ fn verify_collateral_rate(lot_size: u128) -> Result<(), Error> {
     }
 
     Ok(())
-}
-
-pub fn verify(toCKB_data_tuple: &ToCKBCellDataTuple) -> Result<(), Error> {
-    let input_toCKB_data = toCKB_data_tuple
-        .0
-        .as_ref()
-        .expect("inputs contain toCKB cell");
-    let lot_size = match input_toCKB_data.get_xchain_kind() {
-        XChainKind::Btc => input_toCKB_data.get_btc_lot_size()?.get_sudt_amount(),
-        XChainKind::Eth => input_toCKB_data.get_eth_lot_size()?.get_sudt_amount(),
-    };
-    verify_collateral_rate(lot_size)?;
-    verify_burn(lot_size, input_toCKB_data)
 }
