+++ contracts/toCKB-typescript/src/switch/preterm_redeem/mod.rs
@@ -1,14 +1,28 @@
 use crate::switch::ToCKBCellDataTuple;
-use crate::utils::common::{verify_capacity, verify_data};
 use crate::utils::config::SIGNER_FEE_RATE;
-use crate::utils::tools::is_XT_typescript;
+use crate::utils::transaction::is_XT_typescript;
 use crate::utils::types::{Error, ToCKBCellDataView};
+use crate::utils::verifier::{verify_capacity, verify_data};
 use ckb_std::ckb_constants::Source;
 use ckb_std::debug;
 use ckb_std::error::SysError;
 use ckb_std::high_level::{load_cell_data, load_cell_lock, load_cell_lock_hash, load_cell_type};
 use core::result::Result;
 use molecule::prelude::*;
+
+pub fn verify(toCKB_data_tuple: &ToCKBCellDataTuple) -> Result<(), Error> {
+    let input_toCKB_data = toCKB_data_tuple
+        .0
+        .as_ref()
+        .expect("inputs contain toCKB cell");
+    let output_toCKB_data = toCKB_data_tuple
+        .1
+        .as_ref()
+        .expect("outputs contain toCKB cell");
+    verify_capacity()?;
+    let lot_size = verify_data(input_toCKB_data, output_toCKB_data)?;
+    verify_burn(lot_size, output_toCKB_data)
+}
 
 fn verify_burn(lot_size: u128, out_toCKB_data: &ToCKBCellDataView) -> Result<(), Error> {
     let mut deposit_requestor = false;
@@ -26,9 +40,7 @@ fn verify_burn(lot_size: u128, out_toCKB_data: &ToCKBCellDataView) -> Result<(),
             Err(SysError::IndexOutOfBound) => break,
             Err(_err) => panic!("iter input return an error"),
             Ok(cell_type) => {
-                if !(cell_type.is_some()
-                    && is_XT_typescript(&cell_type.unwrap(), lock_hash.as_ref()))
-                {
+                if !is_XT_typescript(&cell_type, lock_hash.as_ref()) {
                     input_index += 1;
                     continue;
                 }
@@ -58,9 +70,7 @@ fn verify_burn(lot_size: u128, out_toCKB_data: &ToCKBCellDataView) -> Result<(),
             Err(SysError::IndexOutOfBound) => break,
             Err(_err) => panic!("iter output return an error"),
             Ok(cell_type) => {
-                if !(cell_type.is_some()
-                    && is_XT_typescript(&cell_type.unwrap(), lock_hash.as_ref()))
-                {
+                if !is_XT_typescript(&cell_type, lock_hash.as_ref()) {
                     output_index += 1;
                     continue;
                 }
@@ -95,18 +105,4 @@ fn verify_burn(lot_size: u128, out_toCKB_data: &ToCKBCellDataView) -> Result<(),
         }
     }
     Ok(())
-}
-
-pub fn verify(toCKB_data_tuple: &ToCKBCellDataTuple) -> Result<(), Error> {
-    let input_toCKB_data = toCKB_data_tuple
-        .0
-        .as_ref()
-        .expect("inputs contain toCKB cell");
-    let output_toCKB_data = toCKB_data_tuple
-        .1
-        .as_ref()
-        .expect("outputs contain toCKB cell");
-    verify_capacity()?;
-    let lot_size = verify_data(input_toCKB_data, output_toCKB_data)?;
-    verify_burn(lot_size, output_toCKB_data)
 }
