+++ contracts/toCKB-typescript/src/utils/mod.rs
@@ -1,5 +1,5 @@
-pub mod common;
-pub mod tools;
+pub mod transaction;
+pub mod verifier;
 
 pub use tockb_types as types;
 pub use tockb_types::config;
